+++ package.json
@@ -7,6 +7,7 @@
     "react": "^16.4.0",
     "react-dom": "^16.4.0",
     "react-google-maps": "^9.4.5",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.4",
     "react-time": "^4.3.0",
     "request": "^2.87.0",
