+++ src/components/App.js
@@ -1,60 +1,33 @@
 import React, { Component } from 'react';
 import Header from './Header/Header.js';
+import home from './home.js';
+import AboutUs from './AboutUs.js';
 import Footer from './Footer/Footer.js';
 import WeatherData from './WeatherCard/WeatherData.js';
+import WeatherCard from './WeatherCard/WeatherCard.js';
 import TrafficCard from './TrafficCard/TrafficCard.js';
+import TrafficCardArchive from './TrafficCardArchive.js';
+import WeatherCardArchive from './WeatherCardArchive.js';
 import CardContainer from './CardContainer/CardContainer.js';
+import {
+  BrowserRouter as Router,
+  Route,
+  Link
+} from 'react-router-dom'
 
 import './App.css';
 
 export default class App extends Component {
-    constructor(props){
-        super(props);
-        this.state = {
-            fixedCard: null
-        };
-    }
-
     render(){
-        let cards = [
-            <WeatherData
-                type='weather'
-                zipcode='22209'
-                key='Weather Card'
-                menuLabel='Weather Card'
-            />,
-            <TrafficCard
-                zoom={12}
-                lat={38.9}
-                lng={-77.08}
-                location='Arlington, VA'
-                description='Current traffic in STSI area'
-                key='Traffic Card'
-                menuLabel='Traffic Card'
-            />
-        ];
-        let cardToDropdown = (card, index) => ({
-            label: card.props.menuLabel,
-            index: index
-        });
         return (
+            <Router>
             <div className='App'>
-                <Header
-                    cardEntries={cards.map(cardToDropdown)}
-                    onChange={(index) => {this.setState({fixedCard: index})}}
-                />
-                <div className='ui centered cards'>
-                    <div className='ui relaxed grid'>
-                        <div className='row' />
-                        <CardContainer
-                            cardDuration={10000}
-                            fadeTime={1000}
-                            fixedCard={this.state.fixedCard}
-                        >{cards}</CardContainer>
-                    </div>
-                </div>
-                <Footer />
+                    <Route path="/" exact component={home}/>
+                    <Route path="/aboutUs" component={AboutUs}/>
+                    <Route path="/WeatherCard" component={WeatherCardArchive}/>
+                    <Route path="/TrafficCard" component={TrafficCardArchive}/>
             </div>
+            </Router>
         );
     }
 }
