+++ src/components/Header/Header.css
@@ -23,8 +23,3 @@ div.ui.item.dropdown:hover {
     background-color: #f25253 !important;
     color: white !important;
 }
-
-.ui.menu .ui.dropdown .menu>.item:hover {
-    background-color: #FFF !important;
-    color: #f25253 !important;
-}
